Reproduce old analysis of response to small/large load increase in vitro
This new notebook reproduces an analysis originally performed in Mathematica. The results were first presented on my SfN 2016 poster.

There are slight, inconsequential discrepancies in the RAUC values and p-values between this Python implementation and the original Mathematica implementation. I suspect this is due to differences in the integration algorithm.

In this version of the notebook, all times for the beginning and end of behaviors are hard-coded for future reproducibility. I will later change the notebook to import behavior times from CSV files not currently under version control. If changes or corrections are made to those files, the notebook will generate different results. This first commit has the times hard-coded so that the original analysis will always be reproducible from this commit.