Replace hard-coded times with time windows in load response notebook
In the previous commit, the start and end times of each behavior in a sequence of three behaviors were hard-coded into the notebook. This commit replaces those times given for individual behaviors with a time window for all behaviors in a sequence. Times for individual behaviors are now imported from CSV annotation files. This has the benefit of keeping that information in a single location that can easily changed or corrected if necessary.

At the time of this commit, these changes have no effect on the results of the notebook. However, because the CSV annotation files are not currently under version control, the results could diverge in the future if corrections or changes are made to the CSV files.