Fix video start time in ephyviewer
For some of my videos, ephyviewer.FrameGrabber.set_file sets the FrameGrabber's start_time to bizarre, seemingly random integers which may be positive or negative. I don't understand the ultimate source of this bug (ephyviewer or PyAV?) nor why it applies to only some videos and not others. However, by simply resetting the start_time to 0 I can fix all cases. This change implements a wrapper for the set_file method which changes start_time to 0 after the method has set it to the buggy value.