Add automated data download capabilities
This commit adds a custom FTP handler that improves upon the standard library's by automatically prompting for username and password when necessary. Functions for downloading files with progress reporting were also added, including DownloadAllDataFiles, which will fetch all files associated with a data set from the remote server, if the appropriate paths are specified.

A download button was added to the MetadataSelector, but for now it is commented out. There is a longstanding bug with ipywidgets that prevents the username and password input fields from appearing (see https://github.com/jupyter-widgets/ipywidgets/issues/44#issuecomment-259968403). This will need to be circumvented when I have more time... For now, manually calling DownloadAllDataFiles(metadata) does the trick.