Update README and dependencies
Thanks to a bug introduced into conda with release 4.6.3 (see conda/conda#8404), instructions for pip fail to be executed by `conda env create` on Windows.

As a workaround, this commit duplicates the pip package list in `requirements.txt` and adds the extra step of running `pip install` to the installation instructions in the README.

Also, neo 0.7.1 or later is needed for compatibility with my experimental branch of elephant.