Change labeling of epochs imported from epoch_encoder_file
`LoadAndPrepareData` reads in the `annotations_file` and `epoch_encoder_file` separately and combines them into a single list, `blk.segments[0].epochs`.

Previously, the thing that distinguished data loaded from the `epoch_encoder_file` was that all items in that CSV file were put into a single Neo Epoch with `name`="Epoch Encoder", and the labels on the items were stored in the Neo Epoch's `labels` list. This contrasted with data loaded from the `annotations_file`, which used the "Type" of each item in the CSV file as a Neo Epoch `name`, and the "Label" of items with matching types to create the Neo Epoch's `labels` list.

This commit changes how epochs loaded from `epoch_encoder_file` are added to `blk.segments[0].epochs` to be more similar to the `annotations_file` behavior. The labels in the `epoch_encoder_file` are used as the `name` of Neo Epochs, and a dummy list of `labels` with value "(from epoch encoder file)" is provided.