Import CausalAlphaKernel from neurotic into utils
CausalAlphaKernel was added to neurotic in version 1.4.0. neurotic avoids dependence on elephant by subclassing from neurotic._elephant_tools.Kernel instead of elephant.kernels.Kernel. Consequently, neurotic's CausalAlphaKernel is not compatible with elephant.statistics.instantaneous_rate because that function tests that the kernel is a subclass of elephant.kernels.Kernel.

To work around this, here a local version of the class is defined in utils.py which inherits from neurotic and from elephant, allowing utils.CausalAlphaKernel to remain compatible with elephant.statistics.instantaneous_rate.